Warn on non-existent SUBSTS.local file, make sure it's 600

+++ dm
@@ -2,6 +2,13 @@
 
 set -e
 
+if [ ! -f SUBSTS.local ]; then
+    echo "Please create the file SUBSTS.local first."
+    exit 1;
+fi
+
+chmod 600 SUBSTS.local
+
 if [ `uname` = "Linux" ]; then
     make --version | grep "GNU Make" > /dev/null 2>&1 || { echo "Please install GNU Make"; exit 1; }
     MAKE=`command -v make`
